Fix raft group force destroy issues
When a raft group loses majority after some followers crash, its leader
can still append new entries. If followers crash after
ApplyRaftGroupMembersCmd is appended to the leader's log, this
membership change will not succeed. This issue causes the metadata group
to retry this membership change repeatedly, because it won't be able to
apply this change successfully.

In this case, the user must manually force-destroy the raft group and any
membership operation on this raft group will be skipped by the metadata
group.