Raft auto-clean up inactive sessions
When a lock() call fails with operation timeout, we don't release the session count on the caller, even if forceUnlock() is unlocked. So it will continue to send heartbeats even if it has not acquired the lock or has no pending lock wait keys, etc. To close such inactive but heart-beating sessions, a clean-up task runs in the background. If a session sends heartbeats for some time but holds no locks / semaphores or has no wait keys, it is automatically closed.