Improve RaftLock and RaftSemaphore idempotency
Results of lock/acquire and unlock/release operations are memorized
per <session, thread>. It is because we know that each <session, thread>
pair will have at most 1 active operation at a time and it will not send
a new operation before it is done with the current operation. Therefore,
we memorize result of the last operation for each <session id, thread>
pair and gc them on session close.